+++ src/components/button-toggle/button-toggle-group/button-toggle-group.test.tsx
@@ -368,6 +368,5 @@ testStyledSystemMargin(
     </ButtonToggleGroup>
   ),
   () => screen.getByTestId("button-toggle-group"),
-  undefined,
   { modifier: "&&&" },
 );