+++ src/components/draggable/draggable.test.tsx
@@ -200,7 +200,6 @@ testStyledSystemMargin(
     </DraggableContainer>
   ),
   () => screen.getByTestId("draggable-container"),
-  undefined,
 );
 
 testStyledSystemPadding(
@@ -212,5 +211,21 @@ testStyledSystemPadding(
     </DraggableContainer>
   ),
   () => screen.getByTestId("draggable-item"),
-  { py: "8px" },
 );
+
+test("should render with default padding when no padding props are passed", () => {
+  render(
+    <DraggableContainer>
+      <DraggableItem id="apple">Apple</DraggableItem>
+    </DraggableContainer>,
+  );
+
+  expect(screen.getByTestId("draggable-item")).toHaveStyleRule(
+    "padding-top",
+    "var(--spacing100)",
+  );
+  expect(screen.getByTestId("draggable-item")).toHaveStyleRule(
+    "padding-bottom",
+    "var(--spacing100)",
+  );
+});