+++ src/components/icon-button/icon-button.test.tsx
@@ -115,6 +115,19 @@ testStyledSystemSpacing(
     </IconButton>
   ),
   () => screen.getByRole("button"),
-  { p: 0 },
   { modifier: "&&" },
 );
+
+test("should render with default padding when no padding props are passed", () => {
+  render(
+    <IconButton onClick={() => {}}>
+      <Icon type="home" />
+    </IconButton>,
+  );
+
+  const iconButton = screen.getByRole("button");
+
+  expect(iconButton).toHaveStyleRule("padding", "var(--spacing000)", {
+    modifier: "&&",
+  });
+});